develop

+++ app/src/main.c
@@ -78,7 +78,7 @@ int main(void)
 
 void assert_failed(uint8_t* file, uint32_t line)
 {
-	SLOG_ERROR("File: %s - Line: %d: asert failed", file, line);
+	SLOG_ERROR("File: %s - Line: %d: asert failed", file, (unsigned int)line);
 	while(1)
 		;
 }
